requirements.txt
langchain
langchain-openai
python-dotenv
langchain-community
mem0ai
fastapi
uvicorn
mcp
fastmcp
requests
asyncio
contextvars 